TI API YAML update
According to the comments by Broadpeak on UserID. The input parameter is renamed as userId. Before it was usersId. The API is indeed modeled to be used for just one user at a time. Improved the documentation to explain how to manage different TI resources for the same Application. The AF invokes the TI API to create many TI resources, for different users or zones. All the resources are aggregated by the applicationId parameter. E.g., to add a new user, the AF creates a new TI Resource, to remove a user, the AF delete that resource. With GET and applicationId as a filter all the related TI resources are returned.