Update describing and harmonising the Edge APIs.md
In the new section "# GSMA OPAG Capgemini APIs ", included the proposed intents and the keywords to tag the APIs. I have not included the strikethrough keywords for many of the APIs as it will require to include most of the keywords other than GMEC as these new intents doesn't look to be added elsewhere.